wiz-agent/requirements.txt
flask>=2.3.0
langchain>=0.1.0
langchain-aws>=0.1.0
langchain-community>=0.0.20
boto3>=1.34.72
botocore>=1.34.72
python-dotenv>=1.0.0
requests>=2.31.0
gunicorn>=21.0.0
pydantic>=1.10.0
